#108 refactor: Switch the API over to a context provider so it's easier to mock
