+++ services/ui/src/powerpi.tsx
@@ -1,6 +1,7 @@
 import ReactDOM from "react-dom";
 import { QueryClient, QueryClientProvider } from "react-query";
 import Site from "./components/Site";
+import { PowerPiAPIContextProvider } from "./hooks/api";
 import "./styles/imports.scss";
 
 const queryClient = new QueryClient();
@@ -8,7 +9,9 @@ const queryClient = new QueryClient();
 const PowerPi = () => {
     return (
         <QueryClientProvider client={queryClient}>
-            <Site />
+            <PowerPiAPIContextProvider>
+                <Site />
+            </PowerPiAPIContextProvider>
         </QueryClientProvider>
     );
 };
